feat(plugins) first plugin (BBS) and plugin bootstrapping
Added a minimal plugin system if I would ever need to add some fun commands that everyone might not want (e.g. a BBS).
The First plugin is a BBS that acts like a way of reading the blog of the site + gives access to a small "guestbook"/message board like in the good old days.